requirements.txt
psycopg2-binary==2.9.9

# 可选依赖（用于可视化）
# 取消注释以启用图形可视化功能
# graphviz==0.20.1

